Project Source Code
This is the source code written in python for integrating it with the streamlit application. Create a new file with designated name in Microsoft VSCode app,  save it with .py extension, load the file, install given modules and libraries in vscode command terminal, create python virtual environment with Python 3.12.2 package installed and run it in the VSCode command terminal that redirects you to show the execution results of the code in localhost page in browser.