Main page logo (optional)
This is also another logo used in the main page of the streamlit web. This logo depicts the machine learning tools used with the artificial intelligence that are essential for performing sales demand forecast and inventory optimization. You can use this logo in the starting position in the main app page or you can use the similar "Main page logo" in the streamlit web page.