Database file
This is the database file used in developing streamlit web for inserting, retrieving and storing data of the users in the streamlit web page. This database file is very important because for performing the user login, signup with username, password and email at the entry of the website page. Save this file in your created project folder for performing above operations using sqlite3 and hashlib libraries.