Animation video.mp4
This is the animation video with .mp4 extension used in the logout page of the dashboard section in the streamlit web page. Download the file and use it in the logout() page by calling the path of the video file from your system into the python source code provided. Observe the results of the video that is inserted in your project code in the streamlit web page.