toml file
This is one of the primary file named .toml file used for the configuration of the streamlit web page. This file consists of buttons colors, display themes, effects and properties. Download the source file, create a folder namely ".streamlit" and  save it as config.toml file inside the .streamlit folder.